tests/submission/TestTWF_Paris/cummings/innerhtml/insert-adjacent.html

<!doctype html>
<!-- Submitted from TestTWF Paris -->
<title>insertAdjacentHTML</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#element {
	display: none;
}
</style>

<div id="element"></div>
<div id="log"></div>

<script>

setup(function () {
	//
});

var testInfo = {
		help:   "http://w3.org/TR/some-specification#truth-and-beauty"
	, assert: [
			"http://dvcs.w3.org/hg/innerhtml/raw-file/tip/index.html#widl-Element-insertAdjacentHTML-void-insertAdjacentHTMLPosition-position-DOMString-text"
		, "http://domparsing.spec.whatwg.org/#insertadjacenthtml()"
	]
	, authors: [
			"Thibaud Cummings <thibaud.cummings@gmail.com>"
		, "David Freyche <davidfreyche@yahoo.fr>"
	]
}

function wrap(text) {
	return '<div><h3>' + text + '</h3></div>';
}

var possiblePositions = {
		'beforebegin': 'previousSibling'
	, 'afterbegin': 'firstChild'
	, 'beforeend': 'lastChild'
	, 'afterend': 'nextSibling'
}

var el = document.querySelector('#element');

Object.keys(possiblePositions).forEach(function(position) {
	var html = wrap(position);
	
	el.insertAdjacentHTML(position, html);
	
	var div = document.createElement('div');
	div.innerHTML = '<h3>' + position + '</h3>';
	
	test(function() {
		assert_equals(el[possiblePositions[position]].innerHTML, div.innerHTML);
	}, 'insertAdjacentHTML(' + position + ', ' + html + ' )');
});

test(function() {
	assert_throws(new TypeError(), function() {
		el.insertAdjacentHTML("giberish", "giberish");
	});
}, "W3C editors spec: Throws a TypeError exception if the position argument has an invalid value.");

test(function() {
	assert_throws('SyntaxError', function() {
		el.insertAdjacentHTML("giberish", "giberish");
	});
}, "WhatWG editors spec: Throws a DOM SyntaxError exception if the position argument has an invalid value.");

test(function() {
	assert_throws("NoModificationAllowedError", function() {
		var root = document.documentElement;
		html.insertAdjacentHTML("beforebegin", "failure");
	});
}, "Throws a DOMException with name NoModificationAllowedError if the given position isn't possible.\n"
	+ "Inserting elements before the root element of a Document.");
	

test(function() {
	assert_throws("NoModificationAllowedError", function() {
		var root = document.documentElement;
		html.insertAdjacentHTML("afterend", "failure");
	});
}, "Throws a DOMException with name NoModificationAllowedError if the given position isn't possible.\n"
	+ "Inserting elements after the end of the root element of a Document.");

</script>
